Set error for unknown command
An associative array of error messages is put on the prototype so that
the error messages may be overridable (e.g. for localization) and also
so that an object does not have to be created for each Ed instance.

An alternative is to have the object as a static property on the
constructor itself, but this does not allow for the errors to be
overridable.